+++ source/tests/test_os_util.py
@@ -12,25 +12,26 @@
 #  or implied. See the License for the specific language governing permissions#
 #  and limitations under the License.                                         #
 ###############################################################################
-from utils import os_util
+from cfct.utils import os_util
 import mock
-from utils.logger import Logger
+from cfct.utils.logger import Logger
 import os
+import pytest
 
 
 log_level = 'info'
 logger = Logger(loglevel=log_level)
 
-
-@mock.patch('utils.os_util.os')
+@pytest.mark.unit
+@mock.patch('cfct.utils.os_util.os')
 def test_make_dir(self, tmpdir):
     os_util.make_dir(tmpdir)
     assert os.path.isdir(tmpdir) is True
     os_util.make_dir(tmpdir, logger)
     assert os.path.isdir(tmpdir) is True
 
-
-@mock.patch('utils.os_util.os')
+@pytest.mark.unit
+@mock.patch('cfct.utils.os_util.os')
 def test_remove_dir(self, tmpdir):
     os_util.remove_dir(tmpdir)
     assert os.path.isdir(tmpdir) is False