+++ source/tests/test_parameter_manipulation.py
@@ -12,8 +12,8 @@
 #  or implied. See the License for the specific language governing permissions#
 #  and limitations under the License.                                         #
 ###############################################################################
-from utils import parameter_manipulation
-
+from cfct.utils import parameter_manipulation
+import pytest
 
 param = {
     "key": "value",
@@ -30,7 +30,7 @@ trans_params = [
     }
 ]
 
-
+@pytest.mark.unit
 def test_transform_params():
     out_params = parameter_manipulation.transform_params(param)
     for idx in range(len(out_params)):
@@ -39,7 +39,7 @@ def test_transform_params():
         assert out_params[idx]['ParameterValue']\
                == trans_params[idx]['ParameterValue']
 
-
+@pytest.mark.unit
 def test_reverse_transform_params():
     rev_param = parameter_manipulation.reverse_transform_params(trans_params)
     for key in rev_param.keys():