+++ source/tests/test_password_generator.py
@@ -12,10 +12,11 @@
 #  KIND, express or implied. See the License for the specific language       #
 #  governing permissions  and limitations under the License.                 #
 ##############################################################################
-from utils import password_generator
+from cfct.utils import password_generator
 import re
+import pytest
 
-
+@pytest.mark.unit
 def test_random_pwd_generator():
     random_pwd_no_additional_string = \
         password_generator.random_pwd_generator(10, 'a')